refactor, to make ownship slicing more accessible
